requirements/production.txt
# Pro-tip: Try not to put anything here. There should be no dependency in
#	production that isn't in development.
-r base.txt



# WSGI Handler
# ------------------------------------------------
gevent==1.0.2
gunicorn==19.3.0

# Static and Media Storage
# ------------------------------------------------
boto==2.38.0
django-storages-redux==1.3


# Mailgun Support
# ---------------
django-mailgun==0.8.0


